Make sure exceptions in view template doesn't cause problems.
If the exception originally was because of e.g. a character encoding
issue (or any other issue related to the request environment data, which
ExceptionNotifer uses in its output mail), then the same error will
surface when generating the output for the exception mail.

This patch rescues any such exceptions, allowing for the exception mail
to reach the developer anyway.